+++ Structurizr.Core.Tests/View/SequenceCounterTests.cs
@@ -7,7 +7,7 @@ namespace Structurizr.Core.Tests
     {
 
         [Fact]
-        public void Test_increment_IncrementsTheCounter_WhenThereIsNoParent()
+        public void Test_increment_IncrementsTheCounter()
         {
             SequenceCounter counter = new SequenceCounter();
             Assert.Equal("0", counter.AsString());
@@ -17,40 +17,6 @@ namespace Structurizr.Core.Tests
 
             counter.Increment();
             Assert.Equal("2", counter.AsString());
-        }
-
-        [Fact]
-        public void Test_counter_WhenThereIsOneParent()
-        {
-            SequenceCounter parent = new SequenceCounter();
-            parent.Increment();
-            Assert.Equal("1", parent.AsString());
-
-            SequenceCounter child = new SequenceCounter(parent);
-            child.Increment();
-            Assert.Equal("1.1", child.AsString());
-
-            child.Increment();
-            Assert.Equal("1.2", child.AsString());
-        }
-
-        [Fact]
-        public void Test_counter_WhenThereAreTwoParents()
-        {
-            SequenceCounter parent = new SequenceCounter();
-            parent.Increment();
-            Assert.Equal("1", parent.AsString());
-
-            SequenceCounter child = new SequenceCounter(parent);
-            child.Increment();
-            Assert.Equal("1.1", child.AsString());
-
-            SequenceCounter grandchild = new SequenceCounter(child);
-            grandchild.Increment();
-            Assert.Equal("1.1.1", grandchild.AsString());
-
-            grandchild.Increment();
-            Assert.Equal("1.1.2", grandchild.AsString());
         }
 
     }