+++ Structurizr.Core/View/SequenceCounter.cs
@@ -23,31 +23,7 @@
 
         public virtual string AsString()
         {
-            if (Parent == null)
-            {
-                return "" + Sequence;
-            }
-            else
-            {
-                return Parent.AsString() + "." + Sequence;
-            }
-        }
-
-        internal SequenceCounter Clone()
-        {
-            SequenceCounter counter;
-            if (Parent == null)
-            {
-                counter = new SequenceCounter();
-            }
-            else
-            {
-                counter = new SequenceCounter(Parent.Clone());
-            }
-
-            counter.Sequence = Sequence;
-
-            return counter;
+            return "" + Sequence;
         }
 
     }