+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Added validation for hex colour codes (on ElementStyle and RelationshipStyle)
 - Removed the "groups" property of documentation sections (__breaking change__).
 - Added support for the HTTP-based health checks feature.
+- Added an ```EndParallelSequence(boolean)``` method to the ```DynamicView``` class, which allows sequence numbering to continue.
 
 ## 0.7.2